fix: change feed component to a reusable component
