+++ src/app/layout.js
@@ -18,7 +18,7 @@ export default function RootLayout({ children }) {
           <div className="w-1/4 border-r border-gray-300">
             <Sidebar />
           </div>
-          <div className="w-1/2">{children} <FeedItem /> </div>
+          <div className="w-1/2">{children}</div>
           <div className="w-1/4 border-l border-gray-300">widgets</div>
         </div>
       </body>
